+++ R/openxlsxCoerce.R
@@ -181,7 +181,7 @@ openxlsxCoerce.coxph <- function(x, rowNames) {
   coef <- x$coefficients
   se <- sqrt(diag(x$var))
   
-  if (is.null(coef) | is.null(se)) {
+  if (is.null(coef) || is.null(se)) {
     stop("Input is not valid")
   }