+++ tests/testthat/test-outlines.R
@@ -8,16 +8,16 @@ test_that("group columns", {
   # Grouping then setting widths updates hidden
   wb <- createWorkbook()
   addWorksheet(wb, "Sheet 1")
-  groupColumns(wb, "Sheet 1", 2:3, hidden = T)
-  setColWidths(wb, "Sheet 1", 2, widths = "18", hidden = F)
+  groupColumns(wb, "Sheet 1", 2:3, hidden = TRUE)
+  setColWidths(wb, "Sheet 1", 2, widths = "18", hidden = FALSE)
 
   expect_equal(attr(wb$colOutlineLevels[[1]], "hidden")[attr(wb$colOutlineLevels[[1]], "names") == 2], "0")
 
   # Setting column widths then grouping
   wb <- createWorkbook()
   addWorksheet(wb, "Sheet 1")
-  setColWidths(wb, "Sheet 1", 2:3, widths = "18", hidden = F)
-  groupColumns(wb, "Sheet 1", 1:2, hidden = T)
+  setColWidths(wb, "Sheet 1", 2:3, widths = "18", hidden = FALSE)
+  groupColumns(wb, "Sheet 1", 1:2, hidden = TRUE)
 
   expect_equal(attr(wb$colWidths[[1]], "hidden")[attr(wb$colWidths[[1]], "names") == 2], "1")
 })
@@ -28,12 +28,40 @@ test_that("group rows", {
   wb <- createWorkbook()
   assign("wb", wb, envir = .GlobalEnv)
   addWorksheet(wb, "Sheet 1")
-  groupRows(wb, "Sheet 1", 1:4, hidden = T)
+  groupRows(wb, "Sheet 1", 1:4, hidden = TRUE)
 
   expect_equal(names(wb$outlineLevels[[1]]), c("1", "2", "3", "4"))
   expect_equal(unique(attr(wb$outlineLevels[[1]], "hidden")), "1")
   rm(wb)
 })
+
+test_that("group rows 2", {
+ 
+  wb <- createWorkbook()
+  addWorksheet(wb, 'Sheet1')
+  addWorksheet(wb, 'Sheet2')
+  writeData(wb, "Sheet1", iris) 
+  writeData(wb, "Sheet2", iris)
+
+  ## create list of groups
+  # lines used for grouping (here: species)
+  grp <- list(
+    seq(2, 51),
+    seq(52, 101),
+    seq(102, 151)
+  )
+  # assign group levels
+  names(grp) <- c("1","0","1") 
+  groupRows(wb, "Sheet1", rows = grp)
+
+  # different grouping
+  names(grp) <- c("1","2","3")
+  groupRows(wb, "Sheet2", rows = grp)
+
+  expect_equal(unique(wb$outlineLevels[[1]]), c("1", "0"))
+  expect_equal(unique(wb$outlineLevels[[2]]), c("1", "2", "3"))
+})
+
 
 
 test_that("ungroup columns", {
@@ -42,8 +70,8 @@ test_that("ungroup columns", {
   # column groupings left
   wb <- createWorkbook()
   addWorksheet(wb, "Sheet 1")
-  setColWidths(wb, "Sheet 1", 2:3, widths = "18", hidden = F)
-  groupColumns(wb, "Sheet 1", 1:3, hidden = T)
+  setColWidths(wb, "Sheet 1", 2:3, widths = "18", hidden = FALSE)
+  groupColumns(wb, "Sheet 1", 1:3, hidden = TRUE)
   ungroupColumns(wb, "Sheet 1", 1:3)
 
   expect_equal(unique(attr(wb$colWidths[[1]], "hidden")[attr(wb$colWidths[[1]], "names") %in% c(2, 3)]), "0")
@@ -53,7 +81,7 @@ test_that("ungroup columns", {
 test_that("ungroup rows", {
   wb <- createWorkbook()
   addWorksheet(wb, "Sheet 1")
-  groupRows(wb, "Sheet 1", 1:3, hidden = T)
+  groupRows(wb, "Sheet 1", 1:3, hidden = TRUE)
   ungroupRows(wb, "Sheet 1", 1:3)
 
   expect_equal(length(wb$outlineLevels[[1]]), 0L)
@@ -105,11 +133,11 @@ test_that("loading workbook preserves outlines", {
   tf <- temp_xlsx("test")
   tf2 <- temp_xlsx("test2")
 
-  saveWorkbook(wbb, tf, overwrite = T)
+  saveWorkbook(wbb, tf, overwrite = TRUE)
   test <- wbb$worksheets[[1]]$copy()
 
   wb <- loadWorkbook(tf)
-  saveWorkbook(wb, tf2, overwrite = T)
+  saveWorkbook(wb, tf2, overwrite = TRUE)
 
   testthat::expect_equal(wb$worksheets[[1]], test)
 
@@ -151,10 +179,10 @@ test_that("Consecutive calls to saveWorkbook doesn't corrupt attributes", {
   tf <- temp_xlsx("test")
   tf2 <- temp_xlsx("test2")
 
-  saveWorkbook(wbb, tf, overwrite = T)
+  saveWorkbook(wbb, tf, overwrite = TRUE)
   test <- wbb$worksheets[[1]]$copy()
 
-  saveWorkbook(wbb, tf2, overwrite = T)
+  saveWorkbook(wbb, tf2, overwrite = TRUE)
 
   testthat::expect_equal(wbb$worksheets[[1]], test)