+++ NEWS.md
@@ -1,5 +1,11 @@
-#openxlsx (development version)
+# openxlsx (development version)
 
+* Fixed warning on `dataValidation(..., type = "list")` ([#342](https://github.com/ycphs/openxlsx/issues/342))
+* Added optional argument to `loadWorkbook` to decide if empty/blank cells should be converted to NA_character_ (the default) or left blank as is
+
+## Improvements
+
+* Improve detectDates ([#288](https://github.com/ycphs/openxlsx/issues/288))
 * Preserve window size and position, also `getWindowSize()` and `setWindowSize()`  ([466](https://github.com/ycphs/openxlsx/pull/466))
 
 # openxlsx 4.2.6
@@ -13,7 +19,7 @@
 
 ## Fixes
 
-* `openxlsx_setOp()` now works with named list ([#215](https://github.com/ycphs/openxlsx/issues/215))  
+* `openxlsx_setOp()` now works with named list ([#215](https://github.com/ycphs/openxlsx/issues/215))
 * `loadWorkbook()` imports `inlineStr`. Values remain `inlineStr` when writing the workbook with `saveWorkbook()`. Similar `read.xlsx` and `readWorkbook` import `inlineStr`.
 * `read.xlsx()` no longer changes random seed ([#183](https://github.com/ycphs/openxlsx/issues/183))
 * fixed a regression that caused fonts to be read in incorrectly ([#207](https://github.com/ycphs/openxlsx/issues/207))
@@ -388,7 +394,7 @@
 *  functions `addFilter` & `removeFilter` to add filters to columns
 
 *  Headers & footers extended, can now be set with `addWorksheet` and `setHeaderFooter`.
-  `setHeader` & `setFooter` deprecated.  
+  `setHeader` & `setFooter` deprecated.
 
 *  "fitToWidth" and "fitToHeight" logicals in `pageSetup`.